Refactored the MangaFile creation code into the DbFactory where we also setup the first LastModified update.
